Update load.py
code where I get the value of the plant to be displayed is wrong.

I instead have to try and in the case of the plantname regardless because I call this function "update_last_scan_plant." in on_preferences_closed. with the old name which may be not a proper plant name.

the previous change here was because I tried to fix this functions behaviour after using the reset button. Said previous change instead made it so that every call of the update_last_scan_plant function call without an argument was equal to using the reset button just before. for the worth of the plant.